vinst

+++ include/camera.h
@@ -8,7 +8,7 @@ typedef struct view View;
 Camera *createCamera(SDL_Rect *pScreenRect);
 int getCamX(Camera *pCamera);
 int getCamY(Camera *pCamera);
-void updateCamera(Camera *pCamera, int targetX, int targetY);
+void updateCamera(Camera *pCamera, int targetX, float targetY);
 void destroyCamera(Camera *pCamera);
 
 #endif
